Split scraper.rb
scraper.rb file had a 6 main classes, let's split them into individual
files and organise them into a gem-like project structure.
