+++ lib/curry/scraper.rb
@@ -1,0 +1,8 @@
+require "curry/scraper/info"
+
+module Scraper
+  USER_AGENTS = ['Mozilla/5.0 (Macintosh; Intel Mac OS X 10_6_8) AppleWebKit/537.13+ (KHTML, like Gecko) Version/5.1.7 Safari/534.57.2',
+                 'Mozilla/5.0 (Windows NT 6.2; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/29.0.1547.2 Safari/537.36',
+                 'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:21.0) Gecko/20130331 Firefox/21.0',
+                 'Mozilla/5.0 (Windows NT 6.3; Trident/7.0; rv:11.0) like Gecko']
+end
